// frontend/src/generated/openapi/model/infoBasicaLibro.ts
/**
 * amoxtli
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Información básica de un libro para ser mostrado en una lista
 */
export interface InfoBasicaLibro { 
    /**
     * ISBN (International Standard Book Number) del libro
     */
    isbn: string;
    /**
     * Autor del libro
     */
    autor: string;
    /**
     * Título del libro
     */
    titulo: string;
    /**
     * URL de la portada del libro
     */
    urlPortada: string;
}